// src/app/page/master/range-waktupinjaman/range-waktupinjaman.interface.ts
import { EventEmitter } from '@angular/core';

export interface RangeWaktuPinjamanComponent {

	alamatEmail?: string;
	faksimile?: string;
	fixedPhone?: string;
	jamBuka?: string;
	jamTutup?: string;
	kdDepartemen?: string;
	kdKelasHead?: string;
	kdModulAplikasi?: string;
	kdPegawaiKepala?: string;
	kode?: string;
	kodeExternal?: string;
	lokasiRange?: string;
	mobilePhone?: string;
	namaExternal?: string;
	namaRange?: string;
	noCounter?: any;
	noRange?: string;
	prefixNoAntrian?: string;
	reportDisplay?: string;
	statusEnabled?: true;
	statusViewData?: true;
	version?: any;
	website?: string;
	noRec?: string;
	label?: any;
	value?: any;

}